+++ app/layout.tsx
@@ -26,5 +26,5 @@ export const metadata: Metadata = {
 };
 
 export default function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
-  return <html lang="en" suppressHydrationWarning><body><Providers>{children}</Providers></body></html>;
+  return <html lang="en" suppressHydrationWarning><body inert suppressHydrationWarning><Providers>{children}</Providers></body></html>;
 }